LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-name-expected.txt


FAIL @container anchored(fallback) matching name assert_equals: expected "yes" but got ""
FAIL @container anchored(fallback) matching name with tactics, but not name assert_equals: expected "yes" but got ""
